bring back changes from djbe
but don't muck with my settings :)

+++ src/Action.m
@@ -7,6 +7,7 @@
 
 #import "Action.h"
 #import "DSLogger.h"
+#import "PrefsWindowController.h"
 
 
 @implementation Action
@@ -176,18 +177,18 @@
 		return nil;
 	if ([appleScriptResult_ descriptorType] != typeAEList)
 		return nil;
-
-	int count = [appleScriptResult_ numberOfItems], i;
+	
+	NSInteger count = [appleScriptResult_ numberOfItems], i;
 	NSMutableArray *list = [NSMutableArray arrayWithCapacity:count];
 	for (i = 1; i <= count; ++i) {		// Careful -- AppleScript lists are 1-based
 		NSAppleEventDescriptor *elt = [appleScriptResult_ descriptorAtIndex:i];
 		if (!elt) {
-			NSLog(@"Oops -- couldn't get descriptor at index %d", i);
+			NSLog(@"Oops -- couldn't get descriptor at index %ld", (long) i);
 			continue;
 		}
 		NSString *val = [elt stringValue];
 		if (!val) {
-			NSLog(@"Oops -- couldn't turn descriptor at index %d into string", i);
+			NSLog(@"Oops -- couldn't turn descriptor at index %ld into string", (long) i);
 			continue;
 		}
 		[list addObject:val];
@@ -317,7 +318,7 @@
 	return NO;
 }
 
-- (int)numberOfItemsInMenu:(NSMenu *)menu
+- (NSUInteger)numberOfItemsInMenu:(NSMenu *)menu
 {
 	return [classes count];
 }
