+++ src/EvidenceSource.m
@@ -8,6 +8,7 @@
 
 #import "DSLogger.h"
 #import "EvidenceSource.h"
+#import "PrefsWindowController.h"
 
 
 @interface EvidenceSource (Private)
@@ -205,7 +206,7 @@
 {
 	if ([dict objectForKey:@"context"]) {
 		// Set up context selector
-		int index = [ruleContext indexOfItemWithRepresentedObject:[dict valueForKey:@"context"]];
+		NSInteger index = [ruleContext indexOfItemWithRepresentedObject:[dict valueForKey:@"context"]];
 		[ruleContext selectItemAtIndex:index];
 	}
 
@@ -262,10 +263,10 @@
 // NSMenu delegates (for adding rules)
 - (BOOL)menu:(NSMenu *)menu updateItem:(NSMenuItem *)item atIndex:(int)index shouldCancel:(BOOL)shouldCancel;
 - (BOOL)menuHasKeyEquivalent:(NSMenu *)menu forEvent:(NSEvent *)event target:(id *)target action:(SEL *)action;
-- (int)numberOfItemsInMenu:(NSMenu *)menu;
+- (NSUInteger)numberOfItemsInMenu:(NSMenu *)menu;
 
 // NSTableViewDataSource protocol methods
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView;
+- (NSUInteger)numberOfRowsInTableView:(NSTableView *)aTableView;
 - (id)tableView:(NSTableView *)aTableView objectValueForTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex;
 - (void)tableView:(NSTableView *)aTableView setObjectValue:(id)anObject forTableColumn:(NSTableColumn *)aTableColumn row:(int)rowIndex;
 
@@ -459,14 +460,14 @@
 	return NO;
 }
 
-- (int)numberOfItemsInMenu:(NSMenu *)menu
+- (NSUInteger)numberOfItemsInMenu:(NSMenu *)menu
 {
 	return [sources count];
 }
 
 #pragma mark NSTableViewDataSource protocol methods
 
-- (int)numberOfRowsInTableView:(NSTableView *)aTableView
+- (NSUInteger)numberOfRowsInTableView:(NSTableView *)aTableView
 {
 	return [sources count];
 }
